+++ src/client/ui/UI.cc
@@ -101,8 +101,8 @@ void UI::draw()
   tf.colour = Mat44::ID;
 
   // set shaders
-  shader.setAmbientLight( Vec4( 0.5f, 0.5f, 0.5f, 1.00f ) );
-  shader.setCaelumLight( ~Vec3( 1.0f, -1.0f, -1.5f ), Vec4( 0.5f, 0.5f, 0.5f, 1.0f ) );
+  shader.setAmbientLight( Vec4( 0.5f, 0.5f, 0.5f, 1.0f ) );
+  shader.setCaelumLight( Vec3( -0.4851f, 0.4851f, 0.7276f ), Vec4( 0.5f, 0.5f, 0.5f, 1.0f ) );
 
   for( int i = 0; i < liber.shaders.length(); ++i ) {
     shader.program( i );