+++ src/tools/openzone.cc
@@ -48,6 +48,7 @@ static void crashHandler()
 #if SDL_MAJOR_VERSION < 2
   SDL_WM_GrabInput( SDL_GRAB_OFF );
 #endif
+  SDL_Quit();
 }
 
 #if defined( __ANDROID__ )